Likes added

+++ FakeNFT/Models/ItemNFT.swift
@@ -13,7 +13,7 @@ struct ItemNFT: Codable, Equatable {
     let images: [String]
     let rating: Int
     let description: String
-    let price: Double
+    let price: Float
     let author: String
     let id: String
 }
